Only trying to remove view root dir if it exists

+++ studip/views.py
@@ -255,7 +255,7 @@ class ViewSynchronizer:
         if directories_to_keep:
             print("The following directories contain unmanaged files and were kept:\n  - "
                     + "\n  - ".join(directories_to_keep))
-        elif self.view.base: # Do not remove root dir
+        elif self.view_dir in directories and self.view.base: # Do not remove root dir
             os.rmdir(self.view_dir)
 
         self.view = None
